kafka-consumer/requirements.txt
kafka-python==2.0.2
pymongo==3.12.0
PyYAML